% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.R
\docType{data}
\name{birth}
\alias{birth}
\title{Birth Data Set}
\format{
A tibble with 33 rows and 4 variables:
\describe{
  \item{ID}{ID corresponding to the mouse, with ID name also including the Sex and Number}
  \item{Sex}{sex of the mouse identified at birth by the lab}
  \item{Num}{the number assigned to the mouse}
  \item{Treatment}{whether the mouse received the treatment, placebo, or other}
}
}
\source{
{James Xenakis Mouse Lab}
}
\usage{
birth
}
\description{
Includes details for the mice gender, body, weight, and identifying number at birth, all assigned and reported by the lab.
}
\keyword{datasets}
