% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_weight.R
\docType{data}
\name{body_weight}
\alias{body_weight}
\title{Body Weight Data Set}
\format{
A tibble with 33 rows and 7 variables:
\describe{
  \item{ID}{ID corresponding to the mouse, with ID name also including the Sex and Number}
  \item{Body Weight 1}{first check of the mouse's body weight}
  \item{Date Body Weight 1}{date at which the mouse's body weight was first checked}
  \item{Body Weight 2}{second check of the mouse's body weight}
  \item{Date Body Weight 2}{date at which the mouse's body weight was checked for the second time}
  \item{Body Weight 3}{third check of the mouse's body weight}
  \item{Date Body Weight 3}{date at which the mouse's body weight was checked for the third time}
}
}
\source{
{James Xenakis Mouse Lab}
}
\usage{
body_weight
}
\description{
Includes measurements for mouse body weights reported during the lab's routine checks.
}
\keyword{datasets}
