% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatter.R
\name{data_formatter}
\alias{data_formatter}
\title{Data Formatter Function}
\usage{
data_formatter(data, columns_to_check)
}
\arguments{
\item{data}{the dataset for the function to act on}

\item{columns_to_check}{the variable/column for the function to act on. the name requiring inspection.}
}
\value{
a gt() table that outputs the rows that have a non-numeric value in any column specified.
}
\description{
Will flag rows that are non-numeric and do not follow the data format that is expected of the column.
}
\examples{
data_formatter(body_weight, c("Body_Weight_1", "Body_Weight_2", "Body_Weight_3"))
}
