% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_comparison.R
\name{join_comparison}
\alias{join_comparison}
\title{Join Comparison Function}
\usage{
join_comparison(df1, df2, join_by)
}
\arguments{
\item{df1}{a data frame of data from timepoint 1}

\item{df2}{a data frame of data from timepoint 2}

\item{join_by}{input of the variable that you would like to join observations by, most likely the ‘ID’ variable}
}
\value{
a gt() table that returns the observations that did not match in the data frame between timepoint 1 and timepoint 2
}
\description{
Revises any rows that do not match based on their identifying number. Intended to flag any change of IDs by researchers.
}
\examples{
join_comparison(df1 = birth, df2 = body_weight, join_by = "ID")
}
