% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\docType{data}
\name{outcome}
\alias{outcome}
\title{Outcome Data Set}
\format{
A tibble with 33 rows and 7 variables:
\describe{
  \item{Subject_ID}{ID corresponding to the mouse, with ID name also including the Sex and Number}
  \item{Outcome 1}{WATCH LECTURE AND UPDATE DOCUMENTATION}
  \item{Date Outcome 1}{WATCH LECTURE AND UPDATE DOCUMENTATION}
  \item{Outcome 2}{WATCH LECTURE AND UPDATE DOCUMENTATION}
  \item{Date Outcome 2}{WATCH LECTURE AND UPDATE DOCUMENTATION}
  \item{Outcome 3}{WATCH LECTURE AND UPDATE DOCUMENTATION}
  \item{Date Outcome 3}{WATCH LECTURE AND UPDATE DOCUMENTATION}
}
}
\source{
{James Xenakis Mouse Lab}
}
\usage{
outcome
}
\description{
Includes details for the WATCH LECTURE AND UPDATE DOCUMENTATION.
}
\keyword{datasets}
