% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detector.R
\name{outlier_detector}
\alias{outlier_detector}
\title{Outlier Detector Function}
\usage{
outlier_detector(data, x)
}
\arguments{
\item{data}{ADD YOUR PARAMETERS HERE AND WHAT THEY SHOULD BE (requirements, data types, etc.)}

\item{x}{}
}
\value{
ADD DESCRIPTION FOR WHAT YOUR FUNCTION RETURNS (what it is, data type, etc.)
}
\description{
Outlier Detector Function
}
\examples{
ADD EXAMPLES OF HOW TO APPLY YOUR FUNCTION
}
