% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_range.R
\name{set_range}
\alias{set_range}
\title{Set Range Function}
\usage{
set_range(data, column_name, min_value, max_value)
}
\arguments{
\item{data}{the dataset for the function to act on}

\item{column_name}{the column for the function to act on, required to be numeric in type but the function handles non-numeric cases as well with a warning.}

\item{min_value}{the minimum numerical value for the range of the column}

\item{max_value}{the maximum numerical value for the range of the column}
}
\value{
2 gt() tables. The first table will show non-numeric values within the dataset to warn the user. The second table draws from the numeric values (the data excluding the non-numeric values) and from observations that fit the specific numeric range set by the user.
}
\description{
Flags any rows that are outside of a set range of values by a numerical column.
}
\examples{
set_range(data=social_dominance, column_name="Q2", min_value=1, max_value=6)
}
