% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typo_flagger.R
\name{typo_flagger}
\alias{typo_flagger}
\title{Identify Treatment Misspelling Function}
\usage{
typo_flagger(data, column, correct_inputs)
}
\arguments{
\item{data}{the dataset for the function to act on}

\item{column}{the variable/column for the function to act on}

\item{correct_inputs}{a vector of correct inputs for the column}
}
\value{
a gt() table that outputs the rows that have incorrect inputs in the specified column
}
\description{
Flags any observations of a categorical column that do not match the specified correct inputs. Observations may be misspelled or another input altogether that has not been updated.
}
\examples{
typo_flagger(data = birth, column = "Treatment", correct_inputs = c("Plac", "Tmt"))
}
