% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\docType{data}
\name{outcome}
\alias{outcome}
\title{Outcome Data Set}
\format{
A tibble with 33 rows and 7 variables:
\describe{
  \item{Subject_ID}{ID corresponding to the mouse, with ID name also including the Sex and Number}
  \item{Outcome 1}{outcome after first check of using Placebo or Treatment on mouse}
  \item{Date Outcome 1}{date of the first check}
  \item{Outcome 2}{outcome after second check of using Placebo or Treatment on mouse}
  \item{Date Outcome 2}{date of the second check}
  \item{Outcome 3}{outcome after third check of using Placebo or Treatment on mouse}
  \item{Date Outcome 3}{date of the third check}
}
}
\source{
{James Xenakis Mouse Lab}
}
\usage{
outcome
}
\description{
Includes details for the outcome of using either the Placebo or the Treatment on the mice, the data these check ins were done by the lab, and the identifier for each mouse.
}
\keyword{datasets}
